+++ src/net-util/export/EndpointAddress.js
@@ -1,7 +1,6 @@
 // Copyright 2022-2024 the Lactoserv Authors (Dan Bornstein et alia).
 // SPDX-License-Identifier: Apache-2.0
 
-import { FormatUtils } from '@this/loggy-intf';
 import { IntfDeconstructable, Sexp } from '@this/sexp';
 import { MustBe } from '@this/typey';
 
@@ -30,7 +29,7 @@ export class EndpointAddress extends IntfDeconstructable {
   #address;
 
   /**
-   * Port number. May be `null` to indicate "unknown."
+   * Port number. May be `null` to indicate "unknown" or "irrelevant."
    *
    * @type {?number}
    */
@@ -48,8 +47,8 @@ export class EndpointAddress extends IntfDeconstructable {
    *
    * @param {?string} address IP address, or `null` if unknown. If non-`null`,
    *   must be a syntactically valid IP address.
-   * @param {?number} portNumber Port number, or `null` if unknown. If
-   *   non-`null`, must be an integer in the range `1..65535`.
+   * @param {?number} portNumber Port number, or `null` if unknown or
+   *   irrelevant. If non-`null`, must be an integer in the range `1..65535`.
    */
   constructor(address, portNumber) {
     super();
@@ -85,15 +84,78 @@ export class EndpointAddress extends IntfDeconstructable {
   /**
    * Gets a friendly string form of this instance. With known address and port,
    * this is the form `<address>:<port>`, with `<address>` bracketed when in
-   * IPv6 form.
+   * IPv6 form (to make it clear where the port number is). Special cases:
+   *
+   * * When {@link #address} is `null`, this returns literally `<unknown>` (with
+   *   the angle brackets) in place of the address.
+   * * When {@link #address} is in the IPv4-in-v6 wrapped form, this returns
+   *   just the IPv4 address (without the `::ffff:` prefix).
+   * * When {@link #portNumber} is `null`, this does not return either the colon
+   *   or any indication of the port.
    *
    * @returns {string} The friendly string form.
    */
   toString() {
     if (!this.#string) {
-      this.#string = FormatUtils.addressPortString(this.#address, this.#portNumber);
+      this.#string = EndpointAddress.endpointString(this.#address, this.#portNumber);
     }
 
     return this.#string;
+  }
+
+
+  //
+  // Static members
+  //
+
+  /**
+   * Makes a human-friendly network address/port string. This is equivalent to
+   * calling `new EndpointAddress(address, portNumber).toString()`, except that
+   * the arguments aren't validated or canonicalized.
+   *
+   * @param {?string} address The address, or `null` if unknown.
+   * @param {?number} [portNumber] The port numer, or `null` if unknown or
+   *   irrelevant.
+   * @returns {string} The friendly string form.
+   */
+  static endpointString(address, portNumber = null) {
+    const portStr = (portNumber === null) ? '' : `:${portNumber}`;
+
+    let addressStr;
+    if (address === null) {
+      // Unknown address.
+      addressStr = '<unknown>';
+    } else if (/:/.test(address)) {
+      // IPv6 form.
+      const wrappedV4 = address.match(/(?<=^\[?::ffff:)(?=.+[.])[^:\]]+(?=\]?$)/)?.[0];
+      if (wrappedV4) {
+        // It's a "wrapped" IPv4 address. Drop the prefix and any brackets.
+        addressStr = wrappedV4;
+      } else if (address.startsWith('[')) {
+        // Already has brackets. Just leave it as-is.
+        addressStr = address;
+      } else {
+        addressStr = `[${address}]`;
+      }
+    } else {
+      // Presumed to be IPv4 form.
+      addressStr = address;
+    }
+
+    return `${addressStr}${portStr}`;
+  }
+
+  /**
+   * Makes a human-friendly network interface specification string. The given
+   * object is expected to either bind `address` and `port` (with `port`
+   * possibly being `null` but _not_ `undefined`), _or_ bind `fd`.
+   *
+   * @param {object} iface The interface specification to convert.
+   * @returns {string} The friendly form.
+   */
+  static networkInterfaceString(iface) {
+    return (Object.hasOwn(iface, 'fd'))
+      ? `/dev/fd/${iface.fd}`
+      : this.endpointString(iface.address, iface.port);
   }
 }